Mode Identity Websocket logic into Challenge verifying screen
